Проект ShareIt. Бронирование вещей. Часть 2 v.3 CheckStyle (спринт 15)

+++ src/main/java/ru/practicum/shareit/item/dao/CommentRepositoryJpa.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import ru.practicum.shareit.item.model.Comment;
 import java.util.List;
 
-public interface CommentRepositoryJpa extends JpaRepository<Comment, Integer> {
+public interface CommentRepositoryJpa extends JpaRepository<Comment, Long> {
     List<Comment> findAllByItemId(Long itemId);
 }
